refactor: 설계 리펙토링
