--- src/main/java/lotto/domain/LottoMachine.java
@@ -1,13 +1,0 @@
-package lotto.domain;
-
-public class LottoMachine {
-
-    private Lottos lottos;
-    private LottoProgram lottoProgram;
-
-
-    public LottoMachine() {
-    }
-
-
-}
