--- src/main/java/lotto/domain/LottoProgram.java
@@ -1,4 +1,0 @@
-package lotto.domain;
-
-public interface LottoProgram {
-}
