--- src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,0 @@
-package lotto.domain;
-
-public interface Lottos {
-}
