+++ src/main/java/lotto/domain/lotto/LottoAnswer.java
@@ -1,27 +1,27 @@
-package lotto.domain.numberlotto.lotto;
+package lotto.domain.lotto;
 
-import lotto.domain.numberlotto.constants.message.InputError;
-import lotto.domain.numberlotto.constants.message.RangeError;
-import lotto.domain.numberlotto.constants.value.LottoRule;
+import lotto.domain.constants.message.InputError;
+import lotto.domain.constants.message.RangeError;
+import lotto.domain.constants.value.LottoRule;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Lotto {
+public class LottoAnswer {
 
-    private final List<Number> numbers;
+    private final List<Number> answerNumbers;
 
-    public Lotto(List<Integer> numbers) {
+    public LottoAnswer(List<Integer> answerNumbers) {
+        validateDuplicate(answerNumbers);
+        validateLength(answerNumbers);
 
-        validateLength(numbers);
-        validateDuplicate(numbers);
-
-        this.numbers = numbers.stream()
+        this.answerNumbers = answerNumbers.stream()
                 .sorted()
                 .map(Number::new)
                 .collect(Collectors.toList());
     }
+
 
     public void validateLength(List<Integer> numbers) {
         if (numbers.size() != LottoRule.COMBINATION_LENGTH.getValue()) {
@@ -30,10 +30,12 @@ public class Lotto {
         }
     }
 
+
     public void validateDuplicate(List<Integer> numbers) {
         if (numbers.size() != new HashSet<>(numbers).size()) {
             System.out.println(InputError.DUPLICATE_LOTTO_NUMBER.getMessage());
             throw new IllegalArgumentException();
         }
     }
+
 }
