+++ src/main/java/lotto/domain/lotto/Lottos.java
@@ -1,0 +1,11 @@
+package lotto.domain.lotto;
+
+import java.util.List;
+
+public class Lottos {
+
+    private List<Lotto> lottos;
+
+
+
+}
