+++ src/main/java/lotto/domain/lottofactory/LottoFactoryImpl.java
@@ -1,7 +1,7 @@
-package lotto.domain.numberlotto.lottofactory;
+package lotto.domain.lottofactory;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import lotto.domain.numberlotto.lotto.Lotto;
+import lotto.domain.lotto.Lotto;
 
 import java.util.List;
