--- src/main/java/lotto/domain/numberlotto/NumberLottoProgram.java
@@ -1,6 +1,0 @@
-package lotto.domain.numberlotto;
-
-import lotto.domain.LottoProgram;
-
-public class NumberLottoProgram implements LottoProgram {
-}
